Updated for 7.2.0.23826 TODO: Update ui for IconFileDataID

+++ SpellWork/DBC/Structures/SpellDurationEntry.cs
@@ -4,6 +4,6 @@
     {
         public int Duration;
         public int MaxDuration;
-        public ushort DurationPerLevel;
+        public int DurationPerLevel;
     }
 }
